Move slow shred multi_fec test to integration tests folder
